% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSpawner.R
\docType{data}
\name{HXB2}
\alias{HXB2}
\title{Pol Region of the HXB2 Strain of HIV}
\format{A string with 1300 Characters}
\usage{
HXB2
}
\description{
Pol Region of the HXB2 Strain of HIV
}
\keyword{datasets}
