% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSpawner.R
\name{spawn_hiv}
\alias{spawn_hiv}
\title{Spawn HIV}
\usage{
spawn_hiv(number, subtype = "B", snps = 100)
}
\arguments{
\item{number}{The number of sequences to spawn}

\item{snps}{The Maximum number of SNPs to insert}
}
\value{
A vector of Mutant HIV Sequences
}
\description{
Spawn HIV
}
