% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSpawner.R
\name{spawn_sequences}
\alias{spawn_sequences}
\title{spawn_sequences}
\usage{
spawn_sequences(n = 20000, snps = 100, seed)
}
\arguments{
\item{n}{The number of sequences to output}

\item{snps}{The maximum number of snps in each mutant sequence}

\item{seed}{The original DNA Sequence from which the outputs will be mutated}
}
\value{
A vector of mutant sequences
}
\description{
spawn_sequences
}
