% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SequenceSpawner.R
\name{write_fasta}
\alias{write_fasta}
\title{write_fasta}
\usage{
write_fasta(sequences, filename)
}
\arguments{
\item{sequences}{A Vector of Sequences}

\item{filename}{A Filename}
}
\description{
write_fasta
}
